+++ fv/conncheck_test.go
@@ -26,12 +26,11 @@ import (
 	. "github.com/onsi/gomega"
 	log "github.com/sirupsen/logrus"
 
-	"github.com/projectcalico/felix/fv/connectivity"
-	"github.com/projectcalico/felix/fv/containers"
-
 	"github.com/projectcalico/libcalico-go/lib/apiconfig"
 
+	"github.com/projectcalico/felix/fv/connectivity"
 	"github.com/projectcalico/felix/fv/infrastructure"
+	"github.com/projectcalico/felix/fv/tcpdump"
 	"github.com/projectcalico/felix/fv/workload"
 )
 
@@ -121,13 +120,13 @@ func describeConnCheckTests(protocol string) bool {
 					})
 
 					It("with tcpdump", func() {
-						tcpdump := containers.AttachTCPDump(felixes[0].Container, "eth0")
-						tcpdump.SetLogEnabled(true)
-						tcpdump.AddMatcher("UDP", regexp.MustCompile(`.*UDP.*`))
-						tcpdump.Start()
+						tcpd := tcpdump.Attach(felixes[0].Container, "eth0")
+						tcpd.SetLogEnabled(true)
+						tcpd.AddMatcher("UDP", regexp.MustCompile(`.*UDP.*`))
+						tcpd.Start()
 						cc.ExpectLoss(felixes[0], hostW[1], 2*time.Second, 20, -1)
 						cc.CheckConnectivityPacketLoss()
-						Eventually(func() int { return tcpdump.MatchCount("UDP") }).Should(BeNumerically(">", 0))
+						Eventually(func() int { return tcpd.MatchCount("UDP") }).Should(BeNumerically(">", 0))
 					})
 				})
 			}
